Implemented authorization on all Contact controllers

+++ controllers/contact.controller.js
@@ -2,30 +2,35 @@ const db = require('../models');
 const Contact = db.contact;
 
 exports.addContact = (req, res) => {
+
+    const userId = req.user.id;
+
     const newContact = {
         firstName,
         lastName,
         birthday
-    } = req.body
+    } = req.body;
 
-    const contact = new Contact(newContact)
+    newContact.userId = userId;
 
-    contact.save(err => {
-        if (err) {
-            res.status(500).json({
-                message: err
-            });
-            return;
-        }
+    new Contact(newContact)
+        .save(err => {
+            if (err) {
+                res.status(500).json({
+                    message: err
+                });
+                return;
+            }
 
-        res.status(200).json({
-            message: "Contact was created successfully!"
+            res.status(200).json({
+                message: "Contact was created successfully!"
+            })
         })
-    })
 }
 
 exports.getOneContact = (req, res) => {
     const contactId = req.params.id;
+    const userId = req.user.id;
 
     Contact
         .findOne({
@@ -46,13 +51,24 @@ exports.getOneContact = (req, res) => {
                 return;
             }
 
-            res.status(200).json(contact)
+            if (contact.userId == userId) {
+                res.status(200).json(contact)
+            } else {
+                res.status(401).json({
+                    message: 'Unauthorized!'
+                });
+                return;
+            }
         })
 }
 
 exports.getAllContacts = (req, res) => {
+    const userId = req.user.id;
+
     Contact
-        .find()
+        .find({
+            userId
+        })
         .exec((err, contacts) => {
             if (err) {
                 res.status(500).json({
@@ -67,6 +83,7 @@ exports.getAllContacts = (req, res) => {
 
 exports.updateContact = (req, res) => {
     const contactId = req.params.id;
+    const userId = req.user.id;
 
     const {
         birthday
@@ -91,32 +108,67 @@ exports.updateContact = (req, res) => {
                 return;
             }
 
-            contact.birthday = birthday;
-            contact.save()
+            if (contact.userId == userId) {
 
-            res.status(200).json({
-                message: "Contact updated successfully!"
-            })
+                contact.birthday = birthday;
+                contact.save()
+
+                res.status(200).json({
+                    message: "Contact updated successfully!"
+                })
+            } else {
+                res.status(401).json({
+                    message: "Unauthorized!"
+                });
+                return;
+            }
 
         })
 }
 
 exports.deleteContact = (req, res) => {
     const contactId = req.params.id;
+    const userId = req.user.id;
 
     Contact
-    .findByIdAndDelete(contactId)
-    .exec((err, contact) => {
-        if(err){
-            res.status(500).json({message: err});
-            return;
-        }
+        .findById(contactId)
+        .exec((err, contact) => {
+            if (err) {
+                res.status(500).json({
+                    message: err
+                });
+                return;
+            }
 
-        if(contact == undefined){
-            res.status(404).json({message: "Contact not found!"});
-            return;
-        }
+            if (contact == undefined) {
+                res.status(404).json({
+                    message: "Contact not found!"
+                });
+                return;
+            }
 
-        res.status(200).json({message: "Contact was deleted successfully!"});
-    })
+            if (contact.userId == userId) {
+
+                Contact
+                    .findByIdAndDelete(contactId)
+                    .exec(err => {
+                        if (err) {
+                            res.status(500).json({
+                                message: err
+                            });
+                            return;
+                        }
+
+                        res.status(200).json({
+                            message: "Contact was deleted successfully!"
+                        });
+                    })
+            } else {
+                res.status(401).json({
+                    message: "Unauthorized!"
+                });
+                return;
+            }
+        })
+
 }
